Remove chainbase locking support
chainbase locking was removed from the develop/1.8 branch over a month ago because it was never used. Well, it was used in one single place: controller_impl::clear_all_undo(), but that was effectively a noop since nothing else used locking. Unfortunately to remove all remnants of boost mutexes (with problematic behavior on newer toolchains) this locking code needs to be removed from 1.7 too. Because this change has been in develop for over a month confidence of no side effects is high.